+++ CHANGE_LOG.md
@@ -1,3 +1,8 @@
+# 1.0.1
+
+* Add helper to test if log exists in redis or has been expired
+* Add helper to get the timestamp of the last written log entry
+
 # 1.0.0
 
 * Initial release